requirements.txt
streamlit
python-dotenv
google-generativeai
together
requests
